REFACTOR ls/f
MOD command option handling
MOD command structure and package hierarchy
MOD command shell and runner
ADD lsf tool --home functionality
MOD consistent debug and verbose
ADD new commands
DEL deprecated funcs and files
DEL deprecated error types
MOD API for log rotator - use more correct & safer scalar types
NOP func and API comments for command package
MOD remove deprecated process control
MOD registrar: use CnC; add os.Signal handling
DEL lsf/system/process/process.go and test
MOD Registry api - expose Stop()
MOD Environment to use updated Registry api
MOD use updated panics.AsyncRecover
ADD migrate command stub

TEST ADHOC OK - tested with existing .lsf portals

+++ schema/port.go
@@ -136,7 +136,7 @@ func NewLocalPort(path string) (*Port, error) {
 	return port, nil
 }
 
-func NewRemotePort(id, host string, portno int) (*Port, error) {
+func NewRemotePort(id, host string, portno uint16) (*Port, error) {
 	portnumStr := fmt.Sprintf("%d", portno)
 	path := fmt.Sprintf("%s:%s", host, portnumStr)
 	address, e := url.Parse(path)
